backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-dotenv
twilio
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
joblib
python-multipart
requests
